BAH-4644 | Refactor. CarbonContainer out of ObsControl as a seperate story file
